fix(bug): fix a major bug which broke the script
The fix involves updating the script to fetch the latest version of the
runners and ensuring the function responsible to do it, outputted the
data aptly. The function was outputting the data as - "v2.319.1" but the
fix involves stripping off the redundant "v". This allows usage of
appropriate formatted strings which would then correlate to the
appropriate download URL for the runners.